refactor(apes): extract ed25519 keygen helpers to lib/keygen.ts
Pulls generateAndSaveKey, readPublicKey, and the ssh-ed25519 wire-format
builder out of commands/enroll.ts so the upcoming `apes agents` namespace
can reuse them without duplicating crypto logic. Also adds an in-memory
variant (generateKeyPairInMemory) for callers that need the keypair as
strings without writing to disk. Behavior of `apes enroll` is unchanged.
